fix: Fix label quick add and potential exceptions when saving label changes

+++ src/server/api/set-labels.ts
@@ -4,7 +4,7 @@
  * Created Date: 2025-12-08 17:43:05
  * Author: 3urobeat
  *
- * Last Modified: 2026-05-06 19:55:37
+ * Last Modified: 2026-05-07 19:49:12
  * Modified By: 3urobeat
  *
  * Copyright (c) 2025 - 2026 3urobeat <https://github.com/3urobeat>
@@ -50,10 +50,10 @@ export default defineEventHandler(async (event): Promise<ApiResponse<void>> => {
 
     // Write to DB
     return await getApiResponse<void>(async () => {
-        await upsertLabelCategories(updatedCategories);
-        await upsertLabels(updatedLabels);
-        await deleteLabels(deletedLabels.flatMap((e) => e.id));
-        await deleteLabelCategories(deletedCategories.flatMap((e) => e.id));
+        if (updatedCategories) await upsertLabelCategories(updatedCategories);
+        if (updatedLabels)     await upsertLabels(updatedLabels);
+        if (deletedLabels)     await deleteLabels(deletedLabels.flatMap((e) => e.id));
+        if (deletedCategories) await deleteLabelCategories(deletedCategories.flatMap((e) => e.id));
     });
 
 });
